fix(sdk-go): Added instructions for caching JSON Web Keys
When verifying session tokens manually, you can fetch the JSON Web Key
for your instance once and cache it until it's rolled.
Updated the verifying-sessions page with a rough outline of the SDK
methods you would need to fetch the JWK and a brief outline of how you
could cache it.
This commit also fixes some links to use a relative path to other docs
pages and the references to the secret API key, so that they are valid
Go strings.