fix(docs): Update satellite domains documentation
Clarifies that users must complete both the sign-in and sign-up flows on the primary domain using Clerk's <SignIn /> component or useSignIn() hook for sign-in and <SignUp /> component or useSignUp() hook for sign-up. It also explains the redirection process for accessing authentication state from a satellite domain, including sign-in and sign-up flows.

Signed-off-by: Kenton Duprey <kenton@clerk.dev>