Fix async code in basic-rbac.mdx
I was using to implement basic RBAC and ended at step 6 and the page was not redirecting when I was not logged in. I found out the checkRole function is async so I need to await it first. 

NOTE: I didn't update the checkRole for step 7 which also needs await from I can tell.