Revert "Update docs/references/nextjs/clerk-middleware.mdx"
This reverts commit 0220928bc9352ab34fbf9584317364e135ea727f.

Decided that we like the personal account highlight. Here's my justification for any future readers who are interested

In general, our section links are quite good - the section title we link to makes it clear that the article is talking about the term of interest.
In this case, we don't have a "personal account" definition to jump to with a section link. As a user, if I was promised the definition of Personal Account and the link went to https://clerk.com/docs/organizations/organization-workspaces#organization-workspaces-in-the-clerk-dashboard,  there's a good chance i'd read the long title and think "ugh, I don't want to read up on the complete theory of organization workspaces in the dashboard right now thank you", and navigate away. I think the text highlight would point me to the part i'm actually interested in!

It's a little bit brittle though - if we restructure the page in the future so that the first appearance of "Personal account" isn't the definition, this will seem broken. Tradeoffs! I think it's worth it in this case, but not a big deal either way.