added username policies
When someone try to create clerk user using backend sdk, passes the username that don't meet the requirements of username plicies, clerk returns a 422 response code that suggests something was wrong with the input data. But there is no clue for the developers that what exactly the main culprit, therefore I have added the instructions to write a username.